Create new MaterialXView module. Module is disabled by default.
