+++ CMakeLists.txt
@@ -15,6 +15,7 @@ option(MATERIALX_BUILD_PYTHON "Build the MaterialX Python package from C++ bindi
 option(MATERIALX_BUILD_DOCS "Create HTML documentation using Doxygen. Requires that Doxygen be installed." OFF)
 option(MATERIALX_INSTALL_PYTHON "Install the MaterialX Python package as a third-party library when the install target is built." ON)
 option(MATERIALX_WARNINGS_AS_ERRORS "Interpret all compiler warnings as errors." OFF)
+option(MATERIALX_BUILD_VIEW "Build the MaterialX View moule. " OFF)
 
 set(MATERIALX_PYTHON_EXECUTABLE "" CACHE FILEPATH
     "Path to the Python executable (e.g. 'C:/Python27/python.exe').")
@@ -32,6 +33,7 @@ mark_as_advanced(MATERIALX_PYTHON_LIBRARY)
 mark_as_advanced(MATERIALX_PYTHON_OCIO_DIR)
 mark_as_advanced(MATERIALX_INSTALL_PYTHON)
 mark_as_advanced(MATERIALX_WARNINGS_AS_ERRORS)
+mark_as_advanced(MATERIALX_BUILD_VIEW)
 
 set(PYTHON_INCLUDE_DIR ${MATERIALX_PYTHON_INCLUDE_DIR})
 set(PYTHON_LIBRARY ${MATERIALX_PYTHON_LIBRARY})
@@ -85,6 +87,10 @@ endif()
 add_subdirectory(source/MaterialXCore)
 add_subdirectory(source/MaterialXFormat)
 add_subdirectory(source/MaterialXShaderGen)
+if (MATERIALX_BUILD_VIEW)
+    add_subdirectory(source/MaterialXView)
+endif(MATERIALX_BUILD_VIEW)
+
 add_subdirectory(source/MaterialXTest)
 
 if(MATERIALX_BUILD_PYTHON)
@@ -95,3 +101,4 @@ endif(MATERIALX_BUILD_PYTHON)
 if(MATERIALX_BUILD_DOCS)
     add_subdirectory(documents)
 endif(MATERIALX_BUILD_DOCS)
+
