+++ source/MaterialXView/Glsl/GLUtil/HardwarePlatform.h
